HireMind-AI/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0
pdfplumber==0.11.0
python-docx==1.1.2
scikit-learn==1.5.0
sentence-transformers==2.7.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==10.3.0
